tools/requirements.txt
# Core dependencies for Atomic Workflow tooling
pyyaml>=6.0
python-ulid>=1.1.0
structlog>=23.1.0
typer>=0.9.0
psycopg2-binary>=2.9.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
